cli/requirements.txt
openai>=1.14.0
python-dotenv>=1.0.1
networkx>=3.2.1
colorama>=0.4.6
